avanzando con la edicion de educacion

+++ src/app/componentes/estudios/estudios.component.ts
@@ -21,9 +21,34 @@ export class EstudiosComponent implements OnInit {
       this.estudios = data
     })
   }
-  public borrar(id: Number): void{
-    this.estudiosService.deleteEducaciones(id).subscribe(data =>{
-      this.leerDatos()
-    })
+
+  agregar(){
+    this.estudios.push(new educacion(0,"","Institucion","Titulo","YYYY","YYYY",1))
+  }
+
+  guardar(estudios: educacion){
+    estudios.nombreinstitucion = document.getElementById("institucion"+estudios.id)!.innerText;
+    estudios.titulo = document.getElementById("titulo"+estudios.id)!.innerText;
+    estudios.fechadesde = (document.getElementById("fecha"+estudios.id)!.innerText).substring(0,4);
+    estudios.fechahasta = (document.getElementById("fecha"+estudios.id)!.innerText).substring(6,11);
+    if(estudios.id==0){
+      this.estudiosService.createEducacion(estudios).subscribe(data =>{
+        this.leerDatos()
+      })
+    }else{
+      this.estudiosService.updateEducacion(estudios).subscribe(data =>{
+        this.leerDatos()
+      })
+    }
+  }
+
+  borrar(id: Number): void{
+    if(id!=0){
+      this.estudiosService.deleteEducaciones(id).subscribe(data =>{
+        this.leerDatos()
+      })
+    }else{
+      this.estudios.pop()
+    }
   }
 }
